separated game manu into another file
